util_buf.c: Only use huge pages when mmap attribute flag is set
When 3d9b9a951 added huge pages support to util_bufs, all pools got
allocated with huge pages when available. This commit allows providers
to control which ones get backed by huge pages by explcitly opting in
with the pool attribute. For instance, this is helpful for cases where a
provider might want to use a pool backed by huge pages for bounce
buffers registered with a NIC but not for buffer pools that hold
unexpected messages.

Co-authored-by: William Zhang <wilzhang@amazon.com>
Signed-off-by: Robert Wespetal <wesper@amazon.com>
Signed-off-by: Raghu Raja <craghun@amazon.com>