common: Work around possible long delay in getaddrinfo()
Previously, in order to convert a name into an address, the name was first
tried as a host name and if failed then tried as an interface name. However,
depending on system configuration, the function getaddrinfo() called in the
first step might incur seconds of delay if the name could not be resolved.
This could result in very slow operation when trying to convert an interface
name into address.

The workaround is to check the input name as an interface name first.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>