prov/rxm: refactor sending rndv ACK
Minor code cleanup. Also use fi_sendmsg with FI_INJECT instead of fi_inject.
If fi_inject fails for whatever reason, the app may fail to notice it. e.g. if
the ACK was the last message that the app sends before it exits, the failure
status may not be conveyed.

Signed-off-by: Arun C Ilango <arun.ilango@intel.com>